Added smart boundary columns.
The idea is that the longest word in the logical column will dictate the width
of the column.  This is as opposed to the current way it works, which is the
word that you are currently on dictating the width of the column.

This is experimental, and there are some problems with the current
implementation.

* I had to using normal mapping instead of visual mapping.
  With vmap, I lose the cursor column by the time the function is called.

* Triggering the function while on whitespace doesn't always give desireable
  results.  I think I'm going to have to make the plugin "whitespace aware".