FIX: Do not extend to lines that are shorter than the start column.
With the cursor on "ab", I get an invalid selection that extends to the "." of the second line (and a beep when smart boundaries are used), instead of the expected selection of lines 3 + 4.

" Use real visual mode mappings to avoid the mapping delay when entering visual
" mode with "v".
xnoremap <silent> ab :<C-u>call TextObjWordBasedColumn("aw")<cr>
xnoremap <silent> aB :<C-u>call TextObjWordBasedColumn("aW")<cr>

This can be fixed by checking that there's actually a character at the start virtual column. The match for the exact column number also ensures that if the position is inside a Tab or in the middle of a double-width character, the selection will not be extended.